+++ src/main/java/com/nhnacademy/illuwa/auth/client/AuthClient.java
@@ -1,6 +1,7 @@
 package com.nhnacademy.illuwa.auth.client;
 
 import com.nhnacademy.illuwa.auth.dto.*;
+import com.nhnacademy.illuwa.auth.dto.payco.SocialLoginRequest;
 import jakarta.validation.Valid;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
