+++ src/main/java/com/nhnacademy/illuwa/auth/client/InactiveVerificationServiceClient.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.illuwa.auth.client;
 
-import com.nhnacademy.illuwa.auth.dto.*;
+import com.nhnacademy.illuwa.auth.dto.message.*;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
