+++ src/main/java/com/nhnacademy/illuwa/auth/controller/LoginController.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.illuwa.auth.controller;
 
 import com.nhnacademy.illuwa.auth.client.AuthClient;
-import com.nhnacademy.illuwa.auth.dto.InactiveCheckResponse;
+import com.nhnacademy.illuwa.auth.dto.message.InactiveCheckResponse;
 import com.nhnacademy.illuwa.auth.dto.MemberLoginRequest;
-import com.nhnacademy.illuwa.auth.dto.SendVerificationRequest;
+import com.nhnacademy.illuwa.auth.dto.message.SendVerificationRequest;
 import com.nhnacademy.illuwa.auth.dto.TokenResponse;
 import com.nhnacademy.illuwa.auth.service.InactiveVerificationService;
 import com.nhnacademy.illuwa.auth.utils.JwtCookieUtil;
@@ -11,7 +11,6 @@ import com.nhnacademy.illuwa.member.enums.Status;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
