+++ src/main/java/com/nhnacademy/illuwa/auth/dto/message/SendMessageResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.auth.dto;
+package com.nhnacademy.illuwa.auth.dto.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class VerifyCodeRequest {
+public class SendMessageResponse {
+    boolean success;
     String email;
-    String code;
+    String message;
+    String text;
 }
