+++ src/main/java/com/nhnacademy/illuwa/auth/dto/message/VerifyCodeRequest.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.auth.dto;
+package com.nhnacademy.illuwa.auth.dto.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SendMessageResponse {
-    boolean success;
+public class VerifyCodeRequest {
     String email;
-    String message;
-    String text;
+    String code;
 }
