+++ src/main/java/com/nhnacademy/illuwa/auth/dto/message/VerifyCodeResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.auth.dto;
+package com.nhnacademy.illuwa.auth.dto.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
