+++ src/main/java/com/nhnacademy/illuwa/auth/dto/payco/SocialLoginRequest.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.auth.dto;
+package com.nhnacademy.illuwa.auth.dto.payco;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
