+++ src/main/java/com/nhnacademy/illuwa/member/client/MemberServiceClient.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.illuwa.member.client;
 
-import com.nhnacademy.illuwa.auth.dto.*;
+import com.nhnacademy.illuwa.auth.dto.payco.PaycoMemberUpdateRequest;
 import com.nhnacademy.illuwa.member.dto.MemberResponse;
 import com.nhnacademy.illuwa.member.dto.MemberUpdateRequest;
 import com.nhnacademy.illuwa.member.dto.PasswordCheckRequest;
